refactor(api): move certification center for admin repository
in src/organizational-entities
